% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{check_lambda_exists}
\alias{check_lambda_exists}
\title{check_lambda_exists}
\usage{
check_lambda_exists(function_name, cred, lambda_server_info)
}
\arguments{
\item{function_name}{a string for the function name}

\item{cred}{a list form of the credentials}

\item{lambda_server_info}{a list form of Lambda server information: id, keys, region}
}
\description{
check if a Lambda function exists
}
\keyword{internal}
