% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_apis.R
\name{.fa_files_root}
\alias{.fa_files_root}
\title{Get local FaaSr files directory}
\usage{
.fa_files_root()
}
\value{
Character string path to the files directory
}
\description{
Internal function to get the files directory for local FaaSr data storage.
Creates the directory if it doesn't exist.
}
\keyword{internal}
