% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_test.R
\name{.faasr_build_adjacency}
\alias{.faasr_build_adjacency}
\title{Build adjacency list from ActionList}
\usage{
.faasr_build_adjacency(action_list)
}
\arguments{
\item{action_list}{List containing the workflow action definitions}
}
\value{
List where each element contains the names of functions that can be invoked next
}
\description{
Internal function to build an adjacency list representation of the workflow
graph from the ActionList using InvokeNext references.
}
\keyword{internal}
