% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{.faasr_get_svc}
\alias{.faasr_get_svc}
\title{.faasr_get_svc}
\usage{
.faasr_get_svc()
}
\value{
object containing all the information
}
\description{
Helper function to get the saved data for 
json, cred, and operations.
}
\examples{
if (interactive()){
svc <- .faasr_get_svc()
}
}
\keyword{internal}
