% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr}
\alias{faasr}
\title{faasr}
\usage{
faasr(json_path = NULL, env_path = NULL)
}
\arguments{
\item{json_path}{a string for the json path}

\item{env_path}{a string for the env(credentials) path}
}
\value{
svc; a set of data consisting of functions and data
}
\description{
FaaSr library client-side main function
It's generating the instance for the user
Users can use the functions with the instance generated by "faasr"
}
