% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_abort_on_multiple_invocations.R
\name{faasr_abort_on_multiple_invocations}
\alias{faasr_abort_on_multiple_invocations}
\title{faasr_abort_on_multiple_invocations}
\usage{
faasr_abort_on_multiple_invocations(faasr, pre)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}

\item{pre}{list with names of functions and corresponding predecessors}
}
\description{
Ensures that only one Action proceeds to execute a User Function if there are multiple triggers
This is necessary because if in the Workflow a function receives multiple triggers, 
multiple Actions are invoked; however, we don't want to execute the same function multiple times.
This function aborts all but the last Action triggered.
}
\keyword{internal}
