% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_lock.R
\name{faasr_anyone_else_interested}
\alias{faasr_anyone_else_interested}
\title{faasr_release}
\usage{
faasr_anyone_else_interested(faasr, target_s3, flag_path, flag_name)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}

\item{target_s3}{the name of target server}

\item{flag_path}{the string value of the path for the flag}

\item{flag_name}{the string value of the name of the flag}
}
\value{
a logical value if there are other nodes
}
\description{
Uses locking algorithm with S3 to enforce single User Function execution when there are multiple predecessors
Checking the flag to find any other nodes are seeking for the lock.
}
\examples{
# This function can be run only in the container
if (interactive()){
check <- faasr_anyone_else_interested(faasr, target_s3, flag_path, flag_name)
}
}
\keyword{internal}
