% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_aws_start_existing_vm}
\alias{faasr_aws_start_existing_vm}
\title{Start existing AWS EC2 instance following FaaSr credential pattern}
\usage{
faasr_aws_start_existing_vm(vm_config, faasr = NULL)
}
\arguments{
\item{vm_config}{VM configuration including InstanceId}

\item{faasr}{FaaSr configuration list}
}
\description{
Start existing AWS EC2 instance following FaaSr credential pattern
}
