% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_aws_start_vm}
\alias{faasr_aws_start_vm}
\title{Start AWS EC2 instance with FaaSr credential pattern}
\usage{
faasr_aws_start_vm(vm_config, faasr = NULL)
}
\arguments{
\item{vm_config}{VM configuration}
}
\description{
Start AWS EC2 instance with FaaSr credential pattern
}
