% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_aws_stop_existing_vm}
\alias{faasr_aws_stop_existing_vm}
\title{Stop existing AWS EC2 instance (replaces faasr_aws_terminate_vm)}
\usage{
faasr_aws_stop_existing_vm(vm_config)
}
\arguments{
\item{vm_config}{VM configuration including InstanceId}
}
\description{
Stop existing AWS EC2 instance (replaces faasr_aws_terminate_vm)
}
