% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_aws_terminate_vm}
\alias{faasr_aws_terminate_vm}
\title{Terminate AWS EC2 instance with FaaSr credential pattern}
\usage{
faasr_aws_terminate_vm(vm_details, vm_config)
}
\arguments{
\item{vm_details}{VM details with instance information}

\item{vm_config}{VM configuration for credentials}
}
\description{
Terminate AWS EC2 instance with FaaSr credential pattern
}
