% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_check_workflow_cycle.R
\name{faasr_check_workflow_cycle}
\alias{faasr_check_workflow_cycle}
\title{faasr_check_workflow_cycle}
\usage{
faasr_check_workflow_cycle(faasr)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}
}
\value{
graph a graph representation of the Workflow
}
\description{
Uses graph depth-first search algorithm to detect cycles
}
\examples{
# this function can be run only inside the container
if (interactive()){
predecessors <- faasr_check_workflow_cycle(faasr)
}
}
\keyword{internal}
