% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_collect_sys_env}
\alias{faasr_collect_sys_env}
\title{faasr_collect_sys_env}
\usage{
faasr_collect_sys_env(faasr, cred)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}
}
\value{
credential list
}
\description{
Collect system environment from the R studio
If the credential name is property set (Servername + Key name),
this function will use Sys.getenv to collect credentials.
If there's already credentials in the list, it skips the procedure.
}
