% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_delete_file.R
\name{faasr_delete_file}
\alias{faasr_delete_file}
\title{faasr_delete_file}
\arguments{
\item{faasr}{list with parsed and validated Payload}

\item{server_name}{string with name of the S3 bucket to use; must match a name declared in the faasr list}

\item{remote_folder}{string with the name of the remote folder where the file is to be deleted from}

\item{remote_file}{string with the name for the file to be deleted}
}
\description{
Helper function to delete a file in an S3 bucket
}
