% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_delete_log}
\alias{faasr_delete_log}
\title{faasr_delete_log}
\usage{
faasr_delete_log(uuid)
}
\arguments{
\item{uuid}{the UUID of the log to retrieve}
}
\value{
nothing is returned
}
\description{
deletes a FaaSr log file stored in S3 
This function deleted a specific FaaSr log file
from the user's S3 FaaSrLog bucket
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
logs <- test$faasr_delete_log('SOME-UUID')
}


}
