% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_execute_regular_function}
\alias{faasr_execute_regular_function}
\title{Execute function without VM orchestration}
\usage{
faasr_execute_regular_function(faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Execute function without VM orchestration
}
