% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_execute_strategy_simple_start_end}
\alias{faasr_execute_strategy_simple_start_end}
\title{Execute Strategy 1: VM start at first function, stop at last function}
\usage{
faasr_execute_strategy_simple_start_end(.faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Execute Strategy 1: VM start at first function, stop at last function
}
