% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_execute_strategy_simple_start_end_fixed}
\alias{faasr_execute_strategy_simple_start_end_fixed}
\title{Execute Strategy 1: Fixed version using existing instance}
\usage{
faasr_execute_strategy_simple_start_end_fixed(.faasr)
}
\arguments{
\item{.faasr}{FaaSr configuration list}
}
\description{
Execute Strategy 1: Fixed version using existing instance
}
