% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_get_file.R
\name{faasr_get_file}
\alias{faasr_get_file}
\title{faasr_get_file}
\arguments{
\item{server_name}{string with name of the S3 bucket to use; must match a name declared in the faasr list}

\item{remote_folder}{string with the name of the remote folder where the file is to be downloaded from}

\item{remote_file}{string with the name for the file to be downloaded from the S3 bucket}

\item{local_folder}{string with the name of the local folder where the file to be downloaded is stored}

\item{local_file}{string with the name of the local file once downloaded}
}
\value{
return nothing / delete the file in the bucket
}
\description{
Helper function to download a file from an S3 bucket to local Action folder
}
\examples{
# This function can be run only in the container
if (interactive()){
faasr_get_file("remote_folder", "remote_file", "local_folder", "local_file")
}
}
