% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_get_folder_list.R
\name{faasr_get_folder_list}
\alias{faasr_get_folder_list}
\title{faasr_get_folder_list}
\arguments{
\item{server_name}{string with name of the S3 bucket to use; must match a name declared in the faasr list}

\item{faasr_prefix}{string with prefix of objects in the S3 bucket}
}
\value{
object_list a list of objects in the bucket
}
\description{
Helper function to get a list of objects in an S3 bucket
}
\examples{
# This function can be run only in the container
if (interactive()){
faasr_get_folder_list(server_name="My_Minio_Bucket")
}
}
