% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_get_function_position_in_workflow}
\alias{faasr_get_function_position_in_workflow}
\title{Determine if function is first, last, or middle in workflow}
\usage{
faasr_get_function_position_in_workflow(.faasr, current_function)
}
\arguments{
\item{current_function}{Name of current function}

\item{faasr}{FaaSr configuration list}
}
\description{
Determine if function is first, last, or middle in workflow
}
