% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_get_log}
\alias{faasr_get_log}
\title{faasr_get_log}
\usage{
faasr_get_log(uuid)
}
\arguments{
\item{uuid}{the UUID of the log to retrieve}
}
\value{
nothing is returned; log is downloaded
}
\description{
downloads a FaaSr log file stored in S3 
This function downloads a specific FaaSr log file
from the user's S3 FaaSrLog bucket to local disk
This can be useful in debugging by helping the user
retrieve the specific log of a FaaSr run
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
logs <- test$faasr_get_log('SOME-UUID')
}


}
