% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_get_log_df}
\alias{faasr_get_log_df}
\title{faasr_get_log_df}
\usage{
faasr_get_log_df()
}
\value{
this function returns nothing
}
\description{
returns a data frame with FaaSr logs stored in S3 
This function queries and aggregates FaaSr log 
information configured with the user's S3 bucket
This can be useful in debugging by helping the user
identify and download a particular log
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
test$faasr_get_log_df()
}
}
