% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_get_user_function_args.R
\name{faasr_get_user_function_args}
\alias{faasr_get_user_function_args}
\title{faasr_get_user_function_args}
\usage{
faasr_get_user_function_args(faasr)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}
}
\value{
args list of User Function arguments
if no args provided, return empty list
}
\description{
Helper function to extract arguments of a User Function from the faasr parsed Payload list
}
\examples{
if (interactive()){
faasr <- list()
faasr$FunctionInvoke <- "F1"
faasr$FunctionList$F1$Arguments$test <- "test"
args <- faasr_get_user_function_args(faasr)
}
}
\keyword{internal}
