% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_get_vm_strategy}
\alias{faasr_get_vm_strategy}
\title{Get VM orchestration strategy from configuration}
\usage{
faasr_get_vm_strategy(.faasr)
}
\arguments{
\item{.faasr}{FaaSr configuration list}
}
\description{
Get VM orchestration strategy from configuration
}
