% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_httr_request}
\alias{faasr_httr_request}
\title{faasr_httr_request}
\usage{
faasr_httr_request(token, url, body = list(), type)
}
\arguments{
\item{token}{a string for the github token}

\item{url}{a string of url}

\item{body}{a list of body}

\item{type}{REST API values; GET/PUT/DELETE/PATCH/POST}
}
\value{
response: response status; 200/202/204/400/etc
}
\description{
the help function to send the curl request to the github
by using the "httr" library.
}
