% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_invoke_workflow}
\alias{faasr_invoke_workflow}
\title{faasr_invoke_workflow}
\usage{
faasr_invoke_workflow(FunctionInvoke = NULL, ...)
}
\arguments{
\item{FunctionInvoke}{a string for invoke function}

\item{...}{a string for underlying functions}
}
\description{
invoke workflow 
This function aggregates the invoke workflow function for
openwhisk, github actions and lambda
This can be used as a cross-platform function
}
