% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_lock.R
\name{faasr_lock}
\alias{faasr_lock}
\alias{faasr_rsm}
\title{faasr_rsm}
\usage{
faasr_rsm(faasr)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}
}
\description{
Uses locking algorithm with S3 to enforce single User Function execution when there are multiple predecessors
}
