% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_log.R
\name{faasr_log}
\alias{faasr_log}
\title{faasr_log}
\arguments{
\item{faasr}{list with parsed and validated Payload}

\item{log_message}{string message to be appended to the log}
}
\description{
Helper function to append to the log file residing in an S3 bucket
the name of the S3 server is implicit from the validated JSON payload, key LoggingServer
the log file is a concatenation of folder "logs" and file name "faasr_log_" + InvocationID + ".txt"
}
