% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_ow_httr_request}
\alias{faasr_ow_httr_request}
\title{faasr_ow_httr_request}
\usage{
faasr_ow_httr_request(
  faasr,
  server,
  action,
  type,
  body = list(),
  ssl = TRUE,
  namespace = NULL
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{server}{a string for the target server}

\item{action}{a string for the target action: /actions, /triggers, /rules}

\item{type}{REST API values; GET/PUT/DELETE/PATCH/POST}

\item{body}{a list of body}

\item{ssl}{SSL CA check; for the SSL certificate: FALSE}

\item{namespace}{a string for the specific namespace e.g., /whisk.system}
}
\value{
an integer value for the response
}
\description{
the help function to send the curl request to the openwhisk
by using the "httr" library.
}
