% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_parse.R
\name{faasr_parse}
\alias{faasr_parse}
\title{faasr_parse}
\usage{
faasr_parse(faasr_payload)
}
\arguments{
\item{faasr_payload}{JSON Payload provided upon Action invocation by the FaaS platform}
}
\value{
faasr list with parsed and validated Payload
}
\description{
This function uses JSON parsing and validation to ensure the Payload is compliant
Two checks are made here: 1) is it a valid JSON format? and 2) does it conform to the FaaSr JSON schema?
If both checks pass, return a list with all the parsed key/value pairs
Otherwise, abort
}
\examples{
# This function can be run only in the container
if (interactive()){
faasr <- faasr_parse(faasr_payload)
}
}
