% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_predecessors_list.R
\name{faasr_predecessors_list}
\alias{faasr_predecessors_list}
\title{faasr_predecessors_list}
\usage{
faasr_predecessors_list(faasr, graph)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}

\item{graph}{graph constructed by the depth first search in faasr_check_workflow_cycle}
}
\value{
a list of "function:predecessors" pairs.
}
\description{
Get the predecessors list, by using the graph from faasr_check_workflow
}
\keyword{internal}
