% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_preview_vm_strategy}
\alias{faasr_preview_vm_strategy}
\title{faasr_preview_vm_strategy}
\usage{
faasr_preview_vm_strategy(vm_strategy = "basic")
}
\arguments{
\item{vm_strategy}{Strategy to use}
}
\value{
Transformed workflow
}
\description{
Preview how VM strategy would transform a workflow without registering it
}
