% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_put_file.R
\name{faasr_put_file}
\alias{faasr_put_file}
\title{faasr_put_file}
\arguments{
\item{server_name}{string with name of the S3 bucket to use; must match a name declared in the faasr list}

\item{local_folder}{string with the name of the local folder where the file to be uploaded resides}

\item{local_file}{string with the name of the local file to be uploaded}

\item{remote_folder}{string with the name of the remote folder where the file is to be uploaded to}

\item{remote_file}{string with the name for the file once uploaded to the S3 bucket}
}
\value{
return nothing / put the file into the bucket
}
\description{
Helper function to upload a file from a local Action folder to an S3 bucket
}
\examples{
# This function can be run only in the container
if (interactive()){
faasr_put_file("local_folder", "local_file", "remote_folder", "remote_file")
}
}
