% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_rank.R
\name{faasr_rank}
\alias{faasr_rank}
\title{faasr_rank}
\usage{
faasr_rank()
}
\value{
return a list of Rank & MaxRank
}
\description{
Helper function to let users identify the rank and max rank in the function
This function is used by users in the container.
}
\examples{
# This function can be run only in the container
if (interactive()){
rank_info <- faasr_rank()
MaxRank <- rank_info$MaxRank 
Rank <- rank_info$Rank
}

}
