% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_google_cloud.R,
%   R/faasr_trigger.R
\name{faasr_refresh_gcp_accesskey}
\alias{faasr_refresh_gcp_accesskey}
\title{faasr_refresh_gcp_accesskey}
\usage{
faasr_refresh_gcp_accesskey(faasr, server)

faasr_refresh_gcp_accesskey(faasr, server)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{server}{a string for the target server}
}
\value{
response faasr object with updated access key for gcp

response faasr object with updated access key for gcp
}
\description{
function to get a new access using the JWT token

function to get a new access using the JWT token
}
\keyword{internal}
