% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_register_workflow}
\alias{faasr_register_workflow}
\title{faasr_register_workflow}
\usage{
faasr_register_workflow(..., vm_strategy = "basic")
}
\arguments{
\item{...}{inputs for timeout, cron, and memory}

\item{vm_strategy}{Strategy to use for resource-intensive functions ("basic" by default)}
}
\value{
return nothing / executes the FaaS
}
\description{
Client tools for registering actions
This aggregates openwhisk, github actions and lambda's
register functions
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
test$register_workflow
}
}
