% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_aws_lambda}
\alias{faasr_register_workflow_aws_lambda}
\title{faasr_register_workflow_aws_lambda}
\usage{
faasr_register_workflow_aws_lambda(faasr, cred, memory = 1024, timeout = 600)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{memory}{an integer for the max size of memory}

\item{timeout}{an integer for the max length of timeout}
}
\description{
starts to register functions in the JSON file for AWS-Lambda
}
\keyword{internal}
