% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_aws_lambda_function_build}
\alias{faasr_register_workflow_aws_lambda_function_build}
\title{faasr_register_workflow_aws_lambda_function_build}
\usage{
faasr_register_workflow_aws_lambda_function_build(
  faasr,
  lambda_function_info,
  function_image_list,
  aws_lambda_role_name,
  cred,
  lambda_server_info,
  memory = 1024,
  timeout = 600
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{lambda_function_info}{a list form of lambda function information: name, actions}

\item{function_image_list}{a list form of lambda function information: name, images}

\item{aws_lambda_role_name}{a string for the lambda role name}

\item{cred}{a list form of the credentials}

\item{lambda_server_info}{a list form of Lambda server information: id, keys, region}

\item{memory}{an integer for the max size of memory}

\item{timeout}{an integer for the max length of timeout}
}
\description{
Create aws lambda functions
}
\keyword{internal}
