% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_aws_lambda_role_create}
\alias{faasr_register_workflow_aws_lambda_role_create}
\title{faasr_register_workflow_aws_lambda_role_create}
\usage{
faasr_register_workflow_aws_lambda_role_create(faasr, cred, lambda_server_info)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{lambda_server_info}{a list form of Lambda server information: id, keys, region}
}
\value{
lambda-role-name a string for the lambda role name
}
\description{
create the aws-lambda role named "faasr-lambda-role"
}
\keyword{internal}
