% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_git_local_repo}
\alias{faasr_register_workflow_git_local_repo}
\title{faasr_register_workflow_git_local_repo}
\usage{
faasr_register_workflow_git_local_repo(repo, ref)
}
\arguments{
\item{repo}{a string for the target repository name}

\item{ref}{a string for the branch of target repository}
}
\description{
Create a local git repository
}
