% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_git_remote_env}
\alias{faasr_register_workflow_git_remote_env}
\title{faasr_register_workflow_git_remote_env}
\usage{
faasr_register_workflow_git_remote_env(repo, cred, token)
}
\arguments{
\item{repo}{a string for the target repository name}

\item{cred}{a list form of the credentials}

\item{token}{a string for the github token}
}
\description{
set env(secrets and variables) to the remote repository
}
\keyword{internal}
