% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_git_remote_repo}
\alias{faasr_register_workflow_git_remote_repo}
\title{faasr_register_workflow_git_remote_repo}
\usage{
faasr_register_workflow_git_remote_repo(token, check, private, repo, ref)
}
\arguments{
\item{token}{a string for the github token}

\item{check}{a logical value whether the remote repository exists}

\item{private}{a logical value to make repository private or not}

\item{repo}{a string for the target repository name}

\item{ref}{a string for the branch of target repository}
}
\value{
a integer value for the result
}
\description{
create / push git remote repository
}
\keyword{internal}
