% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_actions}
\alias{faasr_register_workflow_github_actions}
\title{faasr_register_workflow_github_actions}
\usage{
faasr_register_workflow_github_actions(
  faasr,
  cred,
  cron = NULL,
  runner = FALSE
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{cron}{a string for cron data e.g., */5 * * * *}

\item{runner}{a logical value; enable runner}
}
\description{
register the workflow for github actions.
parse faasr to get the repository list and actions.
create a local/remote repository for the FaaSr actions.
}
