% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_create_dir}
\alias{faasr_register_workflow_github_create_dir}
\title{faasr_register_workflow_github_create_dir}
\usage{
faasr_register_workflow_github_create_dir(server, repo)
}
\arguments{
\item{server}{a string for the server name}

\item{repo}{a string for the target repository name}
}
\description{
Create the directory for the workflows
}
