% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_create_env}
\alias{faasr_register_workflow_github_create_env}
\title{faasr_register_workflow_github_create_env}
\usage{
faasr_register_workflow_github_create_env(server, repo)
}
\arguments{
\item{server}{a string for the server name}

\item{repo}{a string for the target repository name}
}
\description{
Create the environment file .gitignore
}
\keyword{internal}
