% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_create_payload}
\alias{faasr_register_workflow_github_create_payload}
\title{faasr_register_workflow_github_create_payload}
\usage{
faasr_register_workflow_github_create_payload(faasr, repo)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{repo}{a string for the target repository name}
}
\description{
Create the payload JSON file named "payload.json"
}
