% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_create_readme}
\alias{faasr_register_workflow_github_create_readme}
\title{faasr_register_workflow_github_create_readme}
\usage{
faasr_register_workflow_github_create_readme(repo)
}
\arguments{
\item{repo}{a string for the target repository name}
}
\description{
create README.md file for repository description
}
\keyword{internal}
