% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_create_yml_file}
\alias{faasr_register_workflow_github_create_yml_file}
\title{faasr_register_workflow_github_create_yml_file}
\usage{
faasr_register_workflow_github_create_yml_file(
  faasr,
  actionname,
  repo,
  cron = NULL,
  runner = FALSE
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{actionname}{a string for the action name}

\item{repo}{a string for the target repository name}

\item{cron}{a string for cron data e.g., */5 * * * *}

\item{runner}{a logical value; enable runner}
}
\description{
Create a yaml workflow file with the container name
}
\keyword{internal}
