% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_repo_lists}
\alias{faasr_register_workflow_github_repo_lists}
\title{faasr_register_workflow_github_repo_lists}
\usage{
faasr_register_workflow_github_repo_lists(faasr)
}
\arguments{
\item{faasr}{a list form of the JSON file}
}
\value{
a list of "action:server name" pairs.
}
\description{
Parse the faasr and get the list of function:repository
Find actions which is using "Github Actions"
return value's key is action and value is server name.
}
\keyword{internal}
