% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_register_workflow_github_repo_question}
\alias{faasr_register_workflow_github_repo_question}
\title{faasr_register_workflow_github_repo_question}
\usage{
faasr_register_workflow_github_repo_question(check, repo)
}
\arguments{
\item{check}{a logical value}

\item{repo}{a string for the target repository name}
}
\value{
a logical value to make repository private or not
}
\description{
Ask users to update the repository
}
