% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_google_cloud.R
\name{faasr_register_workflow_google_cloud}
\alias{faasr_register_workflow_google_cloud}
\title{faasr_register_workflow_google_cloud}
\usage{
faasr_register_workflow_google_cloud(
  faasr,
  cred,
  ssl = TRUE,
  memory = 1024,
  timeout = 86400,
  storage = NULL
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{ssl}{SSL CA check; for the SSL certificate: FALSE}

\item{memory}{an integer for the max size of memory}

\item{timeout}{an integer for the max length of timeout}
}
\description{
register the workflow for google cloud.
parse faasr to get the server list and actions.
create a actions for the FaaSr actions.
}
\keyword{internal}
