% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_google_cloud.R
\name{faasr_register_workflow_google_cloud_check_exists}
\alias{faasr_register_workflow_google_cloud_check_exists}
\title{faasr_register_workflow_google_cloud_check_exists}
\usage{
faasr_register_workflow_google_cloud_check_exists(
  action,
  server,
  faasr,
  all_jobs
)
}
\arguments{
\item{action}{a string for the target action}

\item{server}{a string for the target server}

\item{faasr}{a list form of the JSON file}

\item{all_jobs}{a list of all existing gcloud jobs}
}
\value{
a logical value; if exists, return TRUE, 
doesn't exist, return FALSE
}
\description{
Check the remote repository is existing on the google cloud
by sending the GET request.
If it exists, return TRUE, doesn't exist, return FALSE
}
\keyword{internal}
