% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_google_cloud.R
\name{faasr_register_workflow_google_cloud_check_user_input}
\alias{faasr_register_workflow_google_cloud_check_user_input}
\title{faasr_register_workflow_google_cloud_check_user_input}
\usage{
faasr_register_workflow_google_cloud_check_user_input(check, actionname)
}
\arguments{
\item{check}{a logical value for target existence}

\item{actionname}{a string for the target action name}
}
\value{
a logical value for overwrite
}
\description{
Ask user input for overwriting the google cloud job
}
\keyword{internal}
