% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_lambda_function_image}
\alias{faasr_register_workflow_lambda_function_image}
\title{faasr_register_workflow_lambda_function_image}
\usage{
faasr_register_workflow_lambda_function_image(faasr, lambda_server_info)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{lambda_server_info}{a list form of Lambda server information: id, keys, region}
}
\value{
function_image_list a list form of lambda function information: name, images
}
\description{
Get aws lambda function image list
}
