% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_lambda_function_lists}
\alias{faasr_register_workflow_lambda_function_lists}
\title{faasr_register_workflow_lambda_function_lists}
\usage{
faasr_register_workflow_lambda_function_lists(faasr)
}
\arguments{
\item{faasr}{a list form of the JSON file}
}
\value{
lambda_function_info a list form of lambda function information: name, actions
}
\description{
Get aws lambda function list
}
\examples{
if (interactive()){
lambda_function_info <- faasr_register_workflow_lambda_function_lists(faasr)
}
}
\keyword{internal}
