% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_lambda_server_info}
\alias{faasr_register_workflow_lambda_server_info}
\title{faasr_register_workflow_lambda_server_info}
\usage{
faasr_register_workflow_lambda_server_info(faasr, cred)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}
}
\value{
lambda_server_info a list form of Lambda server information: id, keys, region
}
\description{
Get aws lambda server information
}
\examples{
if (interactive()){
lambda_server_info <- faasr_register_workflow_lambda_server_info(faasr, cred)
}
}
\keyword{internal}
