% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_register_workflow_openwhisk_action_lists}
\alias{faasr_register_workflow_openwhisk_action_lists}
\title{faasr_register_workflow_openwhisk_action_lists}
\usage{
faasr_register_workflow_openwhisk_action_lists(faasr)
}
\arguments{
\item{faasr}{a list form of the JSON file}
}
\value{
an list of "action name: server name" pairs
}
\description{
Parse the faasr and get the list of function:server
Find actions which is using "OpenWhisk"
return value's key is action and value is server name.
}
\keyword{internal}
