% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_register_workflow_openwhisk_check_exists}
\alias{faasr_register_workflow_openwhisk_check_exists}
\title{faasr_register_workflow_openwhisk_check_exists}
\usage{
faasr_register_workflow_openwhisk_check_exists(ssl, action, server, faasr)
}
\arguments{
\item{ssl}{SSL CA check; for the SSL certificate: FALSE}

\item{action}{a string for the target action: /actions, /triggers, /rules}

\item{server}{a string for the target server}

\item{faasr}{a list form of the JSON file}
}
\value{
a logical value; if exists, return TRUE, 
doesn't exist, return FALSE
}
\description{
Check the remote repository is existing on the openwhisk
by sending the GET request.
If it exists, return TRUE, doesn't exist, return FALSE
}
\examples{
if (interactive()){
check <- faasr_register_workflow_openwhisk_check_exists(ssl, action, server, faasr)
}
}
\keyword{internal}
