% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_register_workflow_openwhisk_check_user_input}
\alias{faasr_register_workflow_openwhisk_check_user_input}
\title{faasr_register_workflow_openwhisk_check_user_input}
\usage{
faasr_register_workflow_openwhisk_check_user_input(check, actionname, type)
}
\arguments{
\item{check}{a logical value for target existence}

\item{actionname}{a string for the target action name}

\item{type}{a string for the action type; actions/triggers/rules}
}
\value{
a logical value for the overwrite
}
\description{
Ask user input for the openwhisk
}
