% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_register_workflow_openwhisk_create_action}
\alias{faasr_register_workflow_openwhisk_create_action}
\title{faasr_register_workflow_openwhisk_create_action}
\usage{
faasr_register_workflow_openwhisk_create_action(
  ssl,
  actionname,
  server,
  faasr,
  memory,
  timeout,
  check
)
}
\arguments{
\item{ssl}{SSL CA check; for the SSL certificate: FALSE}

\item{actionname}{a string for the target action name}

\item{server}{a string for the target server}

\item{faasr}{a list form of the JSON file}

\item{memory}{an integer for the max size of memory}

\item{timeout}{an integer for the max length of timeout}

\item{check}{a logical value for target existence}
}
\description{
Create an action
if it already exists and user wants, update the action
}
