% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_slurm.R
\name{faasr_register_workflow_slurm}
\alias{faasr_register_workflow_slurm}
\alias{\%||\%}
\title{faasr_register_workflow_slurm}
\usage{
x \%||\% y
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{memory}{an integer for the max size of memory in MB}

\item{timeout}{an integer for the max length of timeout in minutes}

\item{cpus}{an integer for the number of CPUs per task}
}
\description{
Register workflow for SLURM cluster.
Parse faasr to get the server list and actions.
Verify connectivity to SLURM REST API endpoints.
}
\keyword{internal}
