% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_slurm.R
\name{faasr_register_workflow_slurm_action_lists}
\alias{faasr_register_workflow_slurm_action_lists}
\title{faasr_register_workflow_slurm_action_lists}
\usage{
faasr_register_workflow_slurm_action_lists(faasr)
}
\arguments{
\item{faasr}{a list form of the JSON file}
}
\value{
a list of "server name: action names" pairs
}
\description{
Parse the faasr and get the list of function:server
Find actions which are using "SLURM"
}
\keyword{internal}
