% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_slurm.R
\name{faasr_register_workflow_slurm_check_connectivity}
\alias{faasr_register_workflow_slurm_check_connectivity}
\title{faasr_register_workflow_slurm_check_connectivity}
\usage{
faasr_register_workflow_slurm_check_connectivity(server, faasr)
}
\arguments{
\item{server}{a string for the target server name}

\item{faasr}{a list form of the JSON file with credentials replaced}
}
\value{
a logical value; if successful, return TRUE, otherwise FALSE
}
\description{
Check the SLURM REST API connectivity by sending a ping request
}
\keyword{internal}
