% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_lock.R
\name{faasr_release}
\alias{faasr_release}
\title{faasr_release}
\usage{
faasr_release(faasr)
}
\arguments{
\item{faasr}{list with parsed and validated Payload}
}
\description{
Uses locking algorithm with S3 to enforce single User Function execution when there are multiple predecessors
Release the lock by deleting the lock file in the bucket
}
\keyword{internal}
