% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_replace_values}
\alias{faasr_replace_values}
\title{faasr_replace_values}
\usage{
faasr_replace_values(faasr, cred)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}
}
\value{
faasr a list form of the JSON file with real keys
}
\description{
replace dummy keys with real keys
Dummy key format: Servername + "_" + Key Name with CAPITAL LETTER
e.g., My_OW_Account + API.key = My_OW_Account_API_KEY
}
