% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_s3_check.R
\name{faasr_s3_check}
\alias{faasr_s3_check}
\title{faasr_s3_check}
\arguments{
\item{faasr}{list with parsed and validated Payload}
}
\value{
faasr list with parsed and validated payload
}
\description{
Check 
1. server's Endpoint&Region and Endpoint has a valid form(http).
2. send a req for the list of buckets to check the status of s3 servers.
3. Check that the bucket name exists.
}
\keyword{internal}
