% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_set_workflow_timer}
\alias{faasr_set_workflow_timer}
\title{faasr_set_workflow_timer}
\usage{
faasr_set_workflow_timer(cron, target = NULL, ...)
}
\arguments{
\item{cron}{a string for cron data e.g., */5 * * * *}

\item{target}{a string for specific function}

\item{...}{a string for underlying functions}
}
\value{
return nothing / set the workflow timer
}
\description{
set cron timer for workflow 
This function aggregates the set cron timer function for
openwhisk, github actions and lambda
This can be used as a cross-platform function
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
test$set_workflow_timer
}
}
