% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_set_workflow_timer_gh}
\alias{faasr_set_workflow_timer_gh}
\title{faasr_set_workflow_timer_gh}
\usage{
faasr_set_workflow_timer_gh(
  faasr,
  cred,
  actionname,
  cron = NULL,
  unset = FALSE
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{actionname}{a string for the target action name}

\item{cron}{a string for cron data e.g., */5 * * * *}

\item{unset}{a logical value; set timer(FALSE) or unset timer(TRUE)}
}
\description{
set/unset cron timer for the github actions.
}
\keyword{internal}
