% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_set_workflow_timer_ld}
\alias{faasr_set_workflow_timer_ld}
\title{faasr_set_workflow_timer_ld}
\usage{
faasr_set_workflow_timer_ld(faasr, cred, target, cron, unset = FALSE)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{target}{a string for the target action}

\item{cron}{a string for cron data e.g., */5 * * * *}

\item{unset}{a logical value; set timer(FALSE) or unset timer(TRUE)}
}
\description{
# set/unset workflow cron timer for lambda
}
