% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_set_workflow_timer_ow}
\alias{faasr_set_workflow_timer_ow}
\title{faasr_set_workflow_timer_ow}
\usage{
faasr_set_workflow_timer_ow(
  faasr,
  cred,
  target,
  cron,
  unset = FALSE,
  ssl = TRUE
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{target}{a string for the target action}

\item{cron}{a string for cron data e.g., */5 * * * *}

\item{unset}{a logical value; set timer(FALSE) or unset timer(TRUE)}

\item{ssl}{SSL CA check; for the SSL certificate: FALSE}
}
\description{
# set/unset workflow cron timer for openwhisk
}
