% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_slurm.R
\name{faasr_slurm_create_job_script}
\alias{faasr_slurm_create_job_script}
\title{faasr_slurm_create_job_script}
\usage{
faasr_slurm_create_job_script(faasr, actionname)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{actionname}{a string for the action name}
}
\value{
string containing the job script
}
\description{
Create SLURM job script for FaaSr execution
}
\keyword{internal}
