% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_slurm.R
\name{faasr_slurm_httr_request}
\alias{faasr_slurm_httr_request}
\title{faasr_slurm_httr_request}
\usage{
faasr_slurm_httr_request(endpoint, method = "GET", headers = NULL, body = NULL)
}
\arguments{
\item{endpoint}{full URL endpoint}

\item{method}{HTTP method}

\item{headers}{HTTP headers}

\item{body}{request body (optional)}
}
\value{
HTTP response object
}
\description{
Helper function to send HTTP requests to SLURM REST API
}
\keyword{internal}
