% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_stop_vm}
\alias{faasr_stop_vm}
\title{faasr_stop_vm}
\usage{
faasr_stop_vm(instance_id = NULL, region = NULL)
}
\arguments{
\item{instance_id}{Optional instance ID override}

\item{region}{Optional region override}
}
\value{
VM information
}
\description{
Manually stop the VM
}
