% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_test.R
\name{faasr_test}
\alias{faasr_test}
\title{faasr_test}
\usage{
faasr_test(
  use_docker = list(use = FALSE, version = docker_default_version, image =
    docker_default_image)
)
}
\arguments{
\item{use_docker}{a list of docker configuration - use, version and image.}
}
\value{
return nothing / executes the FaaS
}
\description{
Client tools for local tests
Users can use local test by using both for Docker and local file system
}
\examples{
if (interactive()){
test <- faasr("test.json", "env")
test$faasr_test()
test$faasr_test(use_docker=list(use=TRUE, version="test", image="docker.io/test/test-docker:image"))
}
}
