% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_unset_workflow_timer}
\alias{faasr_unset_workflow_timer}
\title{faasr_unset_workflow_timer}
\usage{
faasr_unset_workflow_timer(target = NULL, ...)
}
\arguments{
\item{target}{a string for specific function}

\item{...}{a string for underlying functions}
}
\description{
unset cron timer for workflow 
This function aggregates the unset cron timer function for
openwhisk, github actions and lambda
This can be used as a cross-platform function
}
