% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_validate_vm_config_existing}
\alias{faasr_validate_vm_config_existing}
\title{Validate VM configuration for existing instance strategy}
\usage{
faasr_validate_vm_config_existing(vm_config)
}
\arguments{
\item{vm_config}{VM configuration}
}
\description{
Validate VM configuration for existing instance strategy
}
