% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_get_state}
\alias{faasr_vm_get_state}
\title{Retrieve VM state from S3}
\usage{
faasr_vm_get_state(.faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Retrieve VM state from S3
}
