% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_orchestrate}
\alias{faasr_vm_orchestrate}
\title{Main VM orchestration entry point}
\usage{
faasr_vm_orchestrate(.faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Handles VM lifecycle transparently based on strategy
}
