% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_start_simplified}
\alias{faasr_vm_start_simplified}
\title{Start VM instance - simplified version without S3 state}
\usage{
faasr_vm_start_simplified(.faasr)
}
\arguments{
\item{.faasr}{FaaSr configuration list}
}
\description{
Start VM instance - simplified version without S3 state
}
