% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_strategy.R
\name{faasr_vm_strategy}
\alias{faasr_vm_strategy}
\title{faasr_vm_strategy}
\usage{
faasr_vm_strategy(faasr, strategy = "basic")
}
\arguments{
\item{faasr}{list with parsed workflow configuration}

\item{strategy}{string indicating VM strategy to use (default: "basic")}
}
\value{
faasr transformed workflow configuration
}
\description{
Transforms a workflow to use VM-based execution for resource-intensive functions
}
\keyword{internal}
