% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_terminate}
\alias{faasr_vm_terminate}
\title{Terminate VM instance}
\usage{
faasr_vm_terminate(.faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Terminate VM instance
}
