% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_wait_stopped}
\alias{faasr_vm_wait_stopped}
\title{Wait for VM to be fully stopped}
\usage{
faasr_vm_wait_stopped(vm_config)
}
\arguments{
\item{vm_config}{VM configuration including InstanceId}
}
\description{
Wait for VM to be fully stopped
}
