% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_github-actions.R
\name{faasr_workflow_invoke_github}
\alias{faasr_workflow_invoke_github}
\title{faasr_workflow_invoke_github}
\usage{
faasr_workflow_invoke_github(faasr, cred, faas_name, actionname)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{faas_name}{a string for the target server name}

\item{actionname}{a string for the target action name}
}
\description{
invoke the github actions workflow on the github repository.
Async version; does not wait for the result
This will be invoked by faasr_workflow_invoke function
}
