% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_workflow_invoke_lambda}
\alias{faasr_workflow_invoke_lambda}
\title{faasr_workflow_invoke_lambda}
\usage{
faasr_workflow_invoke_lambda(faasr, cred, faas_name, actionname)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{faas_name}{a string for the target server}

\item{actionname}{a string for the target action name}
}
\description{
Invoke a workflow for the lambda
this function is invoked by faasr_workflow_invoke
}
