% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_openwhisk.R
\name{faasr_workflow_invoke_openwhisk}
\alias{faasr_workflow_invoke_openwhisk}
\title{faasr_workflow_invoke_openwhisk}
\usage{
faasr_workflow_invoke_openwhisk(faasr, cred, faas_name, actionname, ssl = TRUE)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{faas_name}{a string for the target server}

\item{actionname}{a string for the target action name}

\item{ssl}{SSL CA check; for the SSL certificate: FALSE}
}
\description{
Invoke a workflow for the openwhisk
this function is invoked by faasr_workflow_invoke
}
