% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_workflow_needs_vm}
\alias{faasr_workflow_needs_vm}
\title{Check if any function in workflow requires VM}
\usage{
faasr_workflow_needs_vm(.faasr)
}
\arguments{
\item{faasr}{FaaSr configuration list}
}
\description{
Check if any function in workflow requires VM
}
