% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_test.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First argument to check}

\item{y}{Default value to return if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Internal operator that returns the second argument if the first is NULL.
}
\keyword{internal}
