% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_arrow_s3_bucket.R
\name{faasr_arrow_s3_bucket}
\alias{faasr_arrow_s3_bucket}
\title{faasr_arrow_s3_bucket}
\arguments{
\item{server_name}{for string, default value is faasr$DefaultDataStore}
}
\value{
s3 representing object for "arrow"
}
\description{
`test` Uses "arrow" library to set up the configurations with given json file and 
provide the object to the users
}
\examples{
# this function can be run only inside the container
if (interactive()){
arrow_s3 <- faasr_arrow_s3_bucket()
arrow_s3$ls
}
}
