% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_tools.R
\name{faasr_register_workflow}
\alias{faasr_register_workflow}
\title{faasr_register_workflow}
\usage{
faasr_register_workflow(...)
}
\arguments{
\item{...}{inputs for timeout, cron, and memory}
}
\description{
Client tools for registering actions
This aggregates openwhisk, github actions and lambda's
register functions
}
