% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_client_api_aws-lambda.R
\name{faasr_register_workflow_lambda_function_info}
\alias{faasr_register_workflow_lambda_function_info}
\title{faasr_register_workflow_lambda_function_info}
\usage{
faasr_register_workflow_lambda_function_info(
  faasr,
  cred,
  lambda_server_info,
  lambda_function_info
)
}
\arguments{
\item{faasr}{a list form of the JSON file}

\item{cred}{a list form of the credentials}

\item{lambda_server_info}{a list form of Lambda server information: id, keys, region}

\item{lambda_function_info}{a list form of lambda function information: name, actions}
}
\value{
lambda_function_info a list form of lambda function information: name, actions, create/update
}
\description{
Get aws lambda function information
}
\examples{
if (interactive()){
faasr <- jsonlite::fromJSON("path/to/json") #from JSON file
cred <- list(keys="creds") #from env file or system environment
server <- list(info="server_info") #from faasr_register_workflow_lambda_server_info()
func_info <- list(info="function_info") #from faasr_register_workflow_lambda_function_lists()
func_info <- faasr_register_workflow_lambda_function_info(faasr, cred, server, func_info)
}
}
\keyword{internal}
