% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faasr_vm_orchestration.R
\name{faasr_vm_put_state}
\alias{faasr_vm_put_state}
\title{Store VM state in S3}
\usage{
faasr_vm_put_state(faasr, vm_details)
}
\arguments{
\item{faasr}{FaaSr configuration list}

\item{vm_details}{VM details to store}
}
\description{
Store VM state in S3
}
